refactor: chunkArray
